Base: implement FileInfo::size()
It was a stub - assert(0) then "return 0" - and since every configuration we
build defines NDEBUG, the assert compiles away and callers just get zero. No
warning, no failure, a plausible-looking answer.

That is why the new Start page reports every file as "0 b", and why
FileBlobManager carries its own fileSize() helper with the comment
"Base::FileInfo::size() is not implemented on this platform, so use Qt". Neither
is a Windows problem; the stub is the whole implementation on every platform.

Implement it with _wstat64 / stat, returning 0 for anything that is not a
regular file, and drop the workaround. The return type widens to uint64_t: the
old unsigned int would have truncated at 4GB, which a CAD project can exceed,
and the only two callers are the two named above.

+++ src/Base/FileInfo.h
@@ -123,8 +123,8 @@ public:
     bool isFile() const;
     /// Checks if it is a directory (not a file)
     bool isDir() const;
-    /// The size of the file
-    unsigned int size() const;
+    /// The size of the file in bytes, or 0 if it does not exist or is not a file.
+    uint64_t size() const;
     /// Returns the time when the file was last modified.
     TimeInfo lastModified() const;
     /// Returns the time when the file was last read (accessed).
